Fix update-script
Amount and size must only be set to amount_backup and size_backup right after the columns
were renamed. If the value-update is executed independently from the renaming-update
all size- and amount-values in data-column will be overwritten with 0 because amount_backup
and size_backup will always be set to 0 after the columns were renamed.