Return form as html string to make sure that the order number exists
The order number is only created once the submit button is clicked.
Because of this, the order number was previously 0 when the html-form
for the submit button was created. To make sure that the order number
can be sent to qenta, the form is now returned as an html string AFTER
the order status has been set to processing and the order number has
been created.